+++ src/event.rs
@@ -133,6 +133,16 @@ pub enum Event<'a, T: 'static> {
   /// This is irreversable - if this event is emitted, it is guaranteed to be the last event that
   /// gets emitted. You generally want to treat this as an "do on quit" event.
   LoopDestroyed,
+
+  /// Emitted when the app is open by external resources, like opening a Url.
+  Opened { event: OpenEvent },
+}
+
+/// What the app is opening.
+#[derive(Debug, PartialEq, Clone)]
+pub enum OpenEvent {
+  /// App is opening an URL.
+  Url(url::Url),
 }
 
 impl<T: Clone> Clone for Event<'static, T> {
@@ -155,6 +165,9 @@ impl<T: Clone> Clone for Event<'static, T> {
       LoopDestroyed => LoopDestroyed,
       Suspended => Suspended,
       Resumed => Resumed,
+      Opened { event } => Opened {
+        event: event.clone(),
+      },
     }
   }
 }
@@ -173,6 +186,7 @@ impl<'a, T> Event<'a, T> {
       LoopDestroyed => Ok(LoopDestroyed),
       Suspended => Ok(Suspended),
       Resumed => Ok(Resumed),
+      Opened { event } => Ok(Opened { event }),
     }
   }
 
@@ -193,6 +207,7 @@ impl<'a, T> Event<'a, T> {
       LoopDestroyed => Some(LoopDestroyed),
       Suspended => Some(Suspended),
       Resumed => Some(Resumed),
+      Opened { event } => Some(Opened { event }),
     }
   }
 }